Revert moving files
