+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardRequestFactory.kt
@@ -2,7 +2,7 @@ package com.paypal.android.cardpayments
 
 import com.paypal.android.corepayments.APIRequest
 import com.paypal.android.corepayments.HttpMethod
-import com.paypal.android.corepayments.api.models.GetOrderRequest
+import com.paypal.android.cardpayments.api.GetOrderRequest
 import org.json.JSONObject
 
 internal class CardRequestFactory {
