+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardResponseParser.kt
@@ -1,15 +1,15 @@
 package com.paypal.android.cardpayments
 
 import com.paypal.android.cardpayments.api.ConfirmPaymentSourceResponse
-import com.paypal.android.corepayments.models.PaymentSource
-import com.paypal.android.corepayments.models.PurchaseUnit
+import com.paypal.android.cardpayments.model.PaymentSource
+import com.paypal.android.cardpayments.model.PurchaseUnit
 import com.paypal.android.corepayments.APIClientError
 import com.paypal.android.corepayments.HttpResponse
 import com.paypal.android.corepayments.OrderErrorDetail
 import com.paypal.android.corepayments.OrderStatus
 import com.paypal.android.corepayments.PayPalSDKError
 import com.paypal.android.corepayments.PaymentsJSON
-import com.paypal.android.corepayments.api.models.GetOrderInfoResponse
+import com.paypal.android.cardpayments.api.GetOrderInfoResponse
 import org.json.JSONException
 
 internal class CardResponseParser {
