+++ CardPayments/src/main/java/com/paypal/android/cardpayments/OrderRequest.kt
@@ -1,6 +1,5 @@
 package com.paypal.android.cardpayments
 
-import com.paypal.android.corepayments.models.OrderIntent
-import com.paypal.android.corepayments.models.PurchaseUnit
+import com.paypal.android.cardpayments.model.PurchaseUnit
 
 data class OrderRequest(val intent: OrderIntent, val purchaseUnits: List<PurchaseUnit>?)
