+++ CardPayments/src/main/java/com/paypal/android/cardpayments/api/ConfirmPaymentSourceResponse.kt
@@ -1,7 +1,7 @@
 package com.paypal.android.cardpayments.api
 
-import com.paypal.android.corepayments.models.PaymentSource
-import com.paypal.android.corepayments.models.PurchaseUnit
+import com.paypal.android.cardpayments.model.PaymentSource
+import com.paypal.android.cardpayments.model.PurchaseUnit
 import com.paypal.android.corepayments.OrderStatus
 
 internal data class ConfirmPaymentSourceResponse(
