+++ CardPayments/src/main/java/com/paypal/android/cardpayments/api/GetOrderRequest.kt
@@ -1,0 +1,3 @@
+package com.paypal.android.cardpayments.api
+
+data class GetOrderRequest(val orderId: String)
