+++ CardPayments/src/main/java/com/paypal/android/cardpayments/model/AuthenticationResult.kt
@@ -1,6 +1,7 @@
-package com.paypal.android.corepayments.models
+package com.paypal.android.cardpayments.model
 
 import com.paypal.android.corepayments.PaymentsJSON
+import com.paypal.android.cardpayments.threedsecure.ThreeDSecureResult
 import org.json.JSONObject
 
 data class AuthenticationResult(
