+++ CardPayments/src/main/java/com/paypal/android/cardpayments/model/PurchaseUnit.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.corepayments.models
+package com.paypal.android.cardpayments.model
 
 import com.paypal.android.corepayments.PaymentsJSON
 
@@ -8,12 +8,12 @@ import com.paypal.android.corepayments.PaymentsJSON
  */
 data class PurchaseUnit(
     val referenceId: String?,
-    val amount: Amount? = null,
+    val amount: com.paypal.android.cardpayments.model.Amount? = null,
     val payee: Payee? = null
 ) {
     constructor(json: PaymentsJSON) : this(
         json.optString("reference_id"),
-        json.optMapObject("amount") { Amount(it) },
+        json.optMapObject("amount") { com.paypal.android.cardpayments.model.Amount(it) },
         json.optMapObject("payee") { Payee(json) }
     )
 }
