+++ Demo/src/main/java/com/paypal/android/ui/card/CardFragment.kt
@@ -17,7 +17,7 @@ import com.paypal.android.cardpayments.ApproveOrderListener
 import com.paypal.android.cardpayments.Card
 import com.paypal.android.cardpayments.CardClient
 import com.paypal.android.cardpayments.CardRequest
-import com.paypal.android.corepayments.models.CardResult
+import com.paypal.android.cardpayments.model.CardResult
 import com.paypal.android.cardpayments.threedsecure.SCA
 import com.paypal.android.corepayments.Address
 import com.paypal.android.corepayments.CoreConfig
